First Run Integration
Integrated FirstSQL into DatabaseConnection. Will run automatically if a database is not found. Will compare sensor names so as to not duplicate in table.
